seperate info drawer code to follow cont and pres pattern
